+++ History.md
@@ -1,3 +1,5 @@
+* SamuelS: WIXFEAT:4543 - Allow Pyro to exclude empty patch transforms.
+
 * BobArnson: WIXBUG:4654 - Add VS14 properties and custom actions. And, as it's a long topic, added anchors and links.
 
 * SeanHall: WIXBUG:4669 - Fix bug in mbahost where it assumed that the CLRCreateInstance function was implemented when it exists.